+++ R/mixscape.R
@@ -127,6 +127,9 @@ CalcPerturbSig <- function(
 #' @param balanced Option to display pathway enrichments for both negative and
 #' positive DE genes.If false, only positive DE gene will be displayed.
 #' @param max.genes Maximum number of genes to use as input to enrichR.
+#' @param logfc.threshold Limit testing to genes which show, on average, at least
+#' X-fold difference (log-scale) between the two groups of cells. Default is 0.25. 
+#' Increasing logfc.threshold speeds up the function, but can miss weaker signals.
 #' @param p.val.cutoff Cutoff to select DE genes.
 #' @param cols A list of colors to use for barplots.
 #' @param enrich.database Database to use from enrichR.
@@ -367,7 +370,9 @@ MixscapeLDA <- function(
 #' @param nt.label Name of non-targeting cell class.
 #' @param npcs Number of principle components to use.
 #' @param verbose Print progress bar.
-#' @inheritParams FindMarkers
+#' @param logfc.threshold Limit testing to genes which show, on average, at least
+#' X-fold difference (log-scale) between the two groups of cells. Default is 0.25. 
+#' Increasing logfc.threshold speeds up the function, but can miss weaker signals.
 #' @return Returns a list of the first 10 PCs from each projection.
 #'
 #' @export
@@ -866,6 +871,9 @@ RunMixscape <- function(
 #' @param max.cells.group Number of cells per identity to plot.
 #' @param max.genes Total number of DE genes to plot.
 #' @param balanced Plot an equal number of genes with both groups of cells.
+#' @param logfc.threshold Limit testing to genes which show, on average, at least
+#' X-fold difference (log-scale) between the two groups of cells. Default is 0.25. 
+#' Increasing logfc.threshold speeds up the function, but can miss weaker signals.
 #' @param order.by.prob Order cells on heatmap based on their mixscape knockout
 #' probability from highest to lowest score.
 #' @param group.by (Deprecated) Option to split densities based on mixscape