+++ tests/testthat/test_preprocessing.R
@@ -473,6 +473,70 @@ test_that("SCTransform v2 works as expected", {
   expect_equal(fa["FCER2", "theta"], Inf)
 })
 
+test_that("SCTransform `vars.to.regress` param works as expected", {
+  # make a copy of the testing data
+  test.data <- object
+  # add a fake mitochondrial gene to the counts matrix
+  counts <- LayerData(test.data, assay = "RNA", layer = "counts")
+  counts <- rbind(counts, 5)
+  rownames(counts)[nrow(counts)] <- "MT-TEST"
+  # use the fake feature to populate a new meta.data column
+  test.data[[ "percent.mt" ]] <- PercentageFeatureSet(
+    test.data,
+    pattern="^MT-"
+  )
+
+  # make sure that `ncells` is smaller than the datset being transformed 
+  # so tha the regression model is trained on a subset of the data - make sure 
+  # the regression is applied to the entire dataset
+  left <- suppressWarnings(
+      SCTransform(
+      test.data,
+      vars.to.regress = NULL,
+      ncells = ncol(test.data) / 2,
+      verbose = FALSE
+    )
+  )
+  right <- suppressWarnings(
+      SCTransform(
+      test.data,
+      vars.to.regress = "percent.mt",
+      ncells = ncol(test.data) / 2,
+      verbose = FALSE
+    )
+  )
+  expect_false(identical(left[["SCT"]]$scale.data, right[["SCT"]]$scale.data))
+
+  # if the `assay` points to an `Assay5` instance the regression is handled
+  # using separate logic
+  test.data[["RNAv5"]] <- CreateAssay5Object(
+    counts = LayerData(
+      test.data,
+      assay = "RNA",
+      layer = "counts"
+    )
+  )
+  left <- suppressWarnings(
+    SCTransform(
+      test.data,
+      assay = "RNAv5",
+      vars.to.regress = NULL,
+      ncells = ncol(test.data) / 2,
+      verbose = FALSE
+    )
+  )
+  right <- suppressWarnings(
+    SCTransform(
+      test.data,
+      assay = "RNAv5",
+      vars.to.regress = "percent.mt",
+      ncells = ncol(test.data) / 2,
+      verbose = FALSE
+    )
+  )
+  expect_false(identical(left[["SCT"]]$scale.data, right[["SCT"]]$scale.data))
+})
+
 test_that("SCTransform is equivalent for BPcells ", {
   skip_on_cran()
   skip_on_cran()